Token Logic implemented and working

+++ tfg-estudio-medico-2019/src/main/java/com/example/tfgestudiomedico2019/repository/UserRepository.java
@@ -1,12 +1,9 @@
 package com.example.tfgestudiomedico2019.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.tfgestudiomedico2019.model.entity.UserEntity;
 
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 	
     UserEntity findByUsername(String username);
 }
